+++ my_list.rb
@@ -1,0 +1,22 @@
+require_relative 'my_enumerable'
+
+class MyList
+  include MyEnumerable
+
+  def initialize(*list)
+    @list = list
+  end
+
+  def each
+    for i in @list do
+      yield i
+    end
+  end
+end
+
+list = MyList.new(1, 2, 3, 4)
+list.all? { |e| e < 5 }
+list.all? { |e| e > 5 }
+list.any? { |e| e == 2 }
+list.any? { |e| e == 5 }
+list.filter(&:even?)
